feat(wasm): split CUE-heavy code into _native/_wasm files across schema and requiredstructure
Move CUE-backed logic out of validate.go, extend.go, rule.go, and
runcache_wiring.go into *_native.go files gated with //go:build !wasm.
Add *_wasm.go stubs for the split-out symbols. Replace sync.Map usage
in runcache.go with a mutex-guarded map so tinygo can compile it.
